* Done ArrayShuffler * Created random locations

+++ Project/Sprint2/src/main/java/com/DragonCard.java
@@ -3,15 +3,15 @@ package main.java.com;
 public class DragonCard {
     private GameCharacters tokenType;
     private int cardCount;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public DragonCard(int cardCount, GameCharacters tokenType) {
         this.tokenType = tokenType;
         this.cardCount = cardCount;
     }
 
-    public void setPos(int x, int y) {
+    public void setPos(float x, float y) {
         this.x = x;
         this.y = y;
     }
